[Refactoring] Padronized components names
