+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 
 import Main from './components/Main';
-import CartEmptyMessage from './components/Cart';
+import CartEmptyMessage from './components/CartEmptyMessage';
 
 import './App.css';
